power: max17050: Fix charge count ext calculation
Bug: 17600205

+++ drivers/power/max17050_gauge.c
@@ -350,7 +350,8 @@ static int max17050_get_charge_ext(struct i2c_client *client, int64_t *batt_char
 		dev_err(&client->dev, "%s: err %d\n", __func__, charge_msb);
 		ret = -EINVAL;
 	} else
-		*batt_charge_ext = ((int16_t) charge_msb <<16 | (int16_t) charge_lsb) * 8;
+		*batt_charge_ext = (int64_t)((int16_t) charge_msb << 16 |
+						(uint16_t) charge_lsb) * 8LL;
 
 	return ret;
 }
